requirements.txt
PyQt5==5.15.0
flask==1.1.2
requests==2.23.0
qtmodern==0.2.0
tensorflow-gpu==2.2.0
imageio==2.9.0
